License header added. ApplicationTest added. Fixed setting the application to the correct startlayer when deepcopying.
