+++ viewer-config-persistence/src/test/java/nl/b3p/viewer/util/TestUtil.java
@@ -1,7 +1,18 @@
 /*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
+ * Copyright (C) 2015 B3Partners B.V.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 package nl.b3p.viewer.util;
 
@@ -16,6 +27,9 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.Persistence;
 import nl.b3p.viewer.config.app.Application;
+import nl.b3p.viewer.config.app.ApplicationLayer;
+import nl.b3p.viewer.config.app.Level;
+import nl.b3p.viewer.config.app.StartLayer;
 import nl.b3p.viewer.config.metadata.Metadata;
 import nl.b3p.viewer.util.databaseupdate.ScriptRunner;
 import org.apache.commons.logging.Log;
@@ -41,6 +55,11 @@ public abstract class TestUtil {
     
     public Long applicationId = 1L;
     public static String originalVersion = null;
+
+    public ApplicationLayer testAppLayer;
+    public Level testLevel;
+    public StartLayer testStartLayer;
+    public Application app;
 
     protected static List<Object> objectsToRemove = new ArrayList<Object>();
 
@@ -160,5 +179,38 @@ public abstract class TestUtil {
                 conn.close();
             }
         }
+    }
+
+    public void initData(boolean deleteAfterwards) {
+        app = new Application();
+        app.setName("testapp");
+        persistEntityTest(app, Application.class, deleteAfterwards);
+
+        testLevel = new Level();
+        testLevel.setName("testStartLayerLevel");
+        app.setRoot(testLevel);
+        entityManager.persist(app);
+        persistEntityTest(testLevel, Level.class, deleteAfterwards);
+
+        testAppLayer = new ApplicationLayer();
+        testAppLayer.setLayerName("testlevel");
+        testLevel.getLayers().add(testAppLayer);
+        persistEntityTest(testAppLayer, ApplicationLayer.class, deleteAfterwards);
+
+        testStartLayer = new StartLayer();
+        testStartLayer.setApplicationLayer(testAppLayer);
+        testStartLayer.setApplication(app);
+        testStartLayer.setSelectedIndex(16);
+        app.getStartLayers().add(testStartLayer);
+
+        testAppLayer.getStartLayers().put(app,testStartLayer);
+
+        entityManager.persist(testAppLayer);
+        entityManager.persist(app);
+
+        persistEntityTest(testStartLayer, StartLayer.class, deleteAfterwards);
+
+        entityManager.getTransaction().commit();
+        entityManager.getTransaction().begin();
     }
 }
